Exercise 4.23

+++ exercise4/controllers/blogs.js
@@ -1,15 +1,19 @@
 const blogsRouter = require("express").Router();
+const userExtractor = require("../utils/middleware").userExtractor;
 const Blog = require("../models/blog");
 const User = require("../models/user");
+const jwt = require("jsonwebtoken");
 
 blogsRouter.get("/", async (request, response) => {
   const blogs = await Blog.find({}).populate("user", {"username": 1, "name": 1, "id": 1});
   response.json(blogs);
 });
 
-blogsRouter.post("/", async (request, response) => {
+
+blogsRouter.post("/", userExtractor, async (request, response) => {
   const body = request.body;
-  const user = await User.findOne({});
+
+  const user = request.user;
 
   const blog = new Blog({
     title: body.title,
@@ -31,9 +35,15 @@ blogsRouter.post("/", async (request, response) => {
   response.status(201).json(savedBlog);
 });
 
-blogsRouter.delete("/:id", async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id);
-  response.status(204).end();
+blogsRouter.delete("/:id", userExtractor, async (request, response) => {
+  const blog = await Blog.findById(request.params.id);
+  const user = request.user;
+  if (blog.user.toString() === user.id) {
+    await Blog.findByIdAndRemove(request.params.id);
+    response.status(204).end();
+  } else {
+    return response.status(401).json({ error: "invalid user" });
+  }
 });
 
 blogsRouter.put("/:id", async (request, response) => {
